refactor(war-poll): make war event poller the sole owner of war state sync
- remove Activity observe loop dependency on `observeClanWar`
- move `WarAttacks` snapshot upsert logic into `WarEventLogService` poll processing
- keep `CurrentWar` and `WarAttacks` updated from the same CoC war fetch in the 5-minute poller
- delete obsolete `WarHistoryService` now that war-state polling no longer writes war data there
